feat: add edit form for course and also implement validations, fix designer namings and also change the type of course credits to CourseCredit enum

+++ Common/Enrollment/EnrollmentProjection.cs
@@ -61,7 +61,7 @@ namespace SystemGroup.General.UniversityManagement.Common
                                     select new
                                     {
                                         EnrollmentID = enrollmentGroup.Key,
-                                        TotalCredits = enrollmentGroup.Sum(c => c.Credits)
+                                        TotalCredits = enrollmentGroup.Sum(c => (int)c.Credits)
                                     };
 
             var enrollmentDetails = from enrollment in inputs
